Title: 숫자 합치기, Time: 58ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241015/숫자 합치기/merge-numbers.cpp
@@ -14,14 +14,14 @@ int main() {
 void solve(){
     cin >> n;
 
-    ll tmp;
-    priority_queue<ll> q;
+    int tmp;
+    priority_queue<int> q;
     for(int i = 0; i < n; i++) cin >> tmp, q.push(-tmp);
 
-    ll ans = 0;
+    int ans = 0;
     while(q.size()>1){
-        ll x = -q.top(); q.pop();
-        ll y = -q.top(); q.pop();
+        int x = -q.top(); q.pop();
+        int y = -q.top(); q.pop();
         ans += x + y;
         q.push(-(x+y));
     }
